fix: 명세서 관련 시간 계산 Double로 수정

+++ BE/src/main/java/com/codaily/project/service/ProjectService.java
@@ -19,7 +19,7 @@ public interface ProjectService {
 
     Project createProject(ProjectCreateRequest request, User user);
 
-    int calculateTotalUserAvailableHours(Long projectId);
+    Double calculateTotalUserAvailableHours(Long projectId);
 
     public FeatureItemReduceResponse reduceFeatureItemsIfNeeded(Long projectId, Long specId);
